[ci] separate outDir for different fixtures
