+++ .github/workflows/ci.yml
@@ -131,10 +131,10 @@ jobs:
             integration-test/*/out*/comp
             integration-test/*/out*/formal
             integration-test/*/out*/sim/report.txt
-            integration-test/*/out/dynamatic_out.txt
-            integration-test/*/out/dynamatic_err.txt
+            integration-test/*/out*/dynamatic_out.txt
+            integration-test/*/out*/dynamatic_err.txt
             integration-test/memory/*/out*/comp
             integration-test/memory/*/out*/sim/report.txt
-            integration-test/memory/*/out/dynamatic_out.txt
-            integration-test/memory/*/out/dynamatic_err.txt
+            integration-test/memory/*/out*/dynamatic_out.txt
+            integration-test/memory/*/out*/dynamatic_err.txt
